% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoExApp.R
\name{CoExApp}
\alias{CoExApp}
\title{run CoExShiny app}
\usage{
CoExApp(absolutefilePath)
}
\arguments{
\item{absolutefilePath}{absolute file path to the directory created by creat_coex_files() command}
}
\value{
opens the shiny web app in a browser
}
\description{
run CoExShiny app
}
