% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_coex_file.R
\name{create_coex_files}
\alias{create_coex_files}
\title{create CoExShiny app}
\usage{
create_coex_files(data, logfc = 1, pval = 0.1, dirname = "CoEx")
}
\arguments{
\item{data}{a seurat object that has been clustered.}

\item{logfc}{numeric value for logfc threshold for markers}

\item{pval}{numeric value for p value threshold for markers}

\item{dirname}{character string for name of the app's directory name}
}
\value{
directory that contains the CoExShiny app and its dependencies
}
\description{
This function takes a suerat object as input and creates a shiny web app. The files contain metadata and a data set containing differentially expressed markers.
}
\examples{
create_coex_files(neural, logfc=1 ,  pval = 0.1, dirname = "CoEx")

}
